Little cleaup
