--- app/src/main/java/com/mrkurilin/sample_project/WidgetViewHodler.kt
@@ -1,8 +1,0 @@
-package com.mrkurilin.sample_project
-
-import android.view.View
-import android.view.ViewStub
-import android.widget.TextView
-import androidx.core.view.isVisible
-import androidx.recyclerview.widget.RecyclerView
-
